refactor(pdf): extract private EPUBCheck adapter
